# Add transceiver pm CLI for advanced window-based performance monitoring