--- a/scripts/sfpshow
+++ b/scripts/sfpshow
@@ -10,7 +10,9 @@ import ast
 import os
 import re
 import sys
+import time
 from typing import Dict
+from datetime import datetime
 
 import click
 from natsort import natsorted
@@ -262,7 +264,12 @@ def display_invalid_intf_status(intf_name):
     click.echo(output)
 
 class SFPShow(object):
-    def __init__(self, intf_name, namespace_option, dump_dom=False):
+    PM_WINDOW_NUM_IN_DB = {'60sec': 15, '15min': 12, '24hrs': 2}
+    PM_WINDOW_START_INDEX_IN_DB = {'60sec': 1, '15min': 16, '24hrs': 28}
+    PM_TABLE_NAME_IN_DB = 'TRANSCEIVER_PM_WINDOW_STATS'
+
+    def __init__(self, intf_name, namespace_option, dump_dom=False,
+                 pm_window_size=None, pm_window_index=None):
         super(SFPShow, self).__init__()
         self.db = None
         self.intf_name = intf_name
@@ -272,6 +279,8 @@ class SFPShow(object):
         self.intf_pm: Dict[str, str] = {}
         self.intf_status: Dict[str, str] = {}
         self.multi_asic = multi_asic_util.MultiAsic(namespace_option=namespace_option)
+        self.pm_window_size = pm_window_size
+        self.pm_window_index = pm_window_index
 
     # Convert dict values to cli output string
     def format_dict_value_to_string(self, sorted_key_table,
@@ -495,11 +504,49 @@ class SFPShow(object):
         else:
             return str(field)
 
-    def convert_interface_sfp_pm_to_cli_output_string(self, state_db, interface_name):
-        sfp_pm_dict = state_db.get_all(
-            self.db.STATE_DB, 'TRANSCEIVER_PM|{}'.format(interface_name))
-        sfp_threshold_dict = state_db.get_all(
-            state_db.STATE_DB, 'TRANSCEIVER_DOM_THRESHOLD|{}'.format(interface_name))
+    def beautify_pm_timestamp(self, secs_since_epoch):
+        if secs_since_epoch is None:
+            return 'N/A'
+        dt_object = datetime.utcfromtimestamp(secs_since_epoch)
+        formatted_time = dt_object.strftime("%a %b %d %H:%M:%S UTC %Y")
+        return formatted_time
+
+    def get_pm_dict_for_specified_window(self, db, interface_name,
+                                         window_size, window_index):
+        pm_windows_dict = db.get_all(
+            db.STATE_DB, '{}|{}'.format(self.PM_TABLE_NAME_IN_DB, interface_name))
+        if not pm_windows_dict:
+            return None
+        window_num = self.PM_WINDOW_NUM_IN_DB[window_size]
+        start_win_idx_in_db = self.PM_WINDOW_START_INDEX_IN_DB[window_size]
+        end_win_idx_in_db = start_win_idx_in_db + window_num - 1
+        pm_windows_for_win_size = []
+        cur_win_idx = -1
+        for i in range(start_win_idx_in_db, end_win_idx_in_db + 1):
+            win_name = 'window' + str(i)
+            if (win_name not in pm_windows_dict or
+                pm_windows_dict[win_name] == 'N/A'):
+                pm_dict = None
+            else:
+                pm_dict = ast.literal_eval(pm_windows_dict[win_name])
+            pm_windows_for_win_size.append(pm_dict)
+            if (pm_dict is not None and
+                'pm_win_current' in pm_dict and
+                pm_dict['pm_win_current'] == 'true'):
+                cur_win_idx = i
+        if cur_win_idx == -1:
+            return None
+        return pm_windows_for_win_size[(cur_win_idx -
+                                        start_win_idx_in_db -
+                                        window_index +
+                                        window_num) % window_num]
+
+    def convert_interface_sfp_pm_to_cli_output_string(self, db, interface_name,
+                                                      window_size, window_index):
+        sfp_pm_dict = self.get_pm_dict_for_specified_window(
+            db, interface_name, window_size, window_index)
+        sfp_threshold_dict = db.get_all(
+            db.STATE_DB, 'TRANSCEIVER_DOM_THRESHOLD|{}'.format(interface_name))
         table = []
         indent_num = 4
         indent = ' ' * indent_num
@@ -517,8 +564,10 @@ class SFPShow(object):
                 for suffix in ZR_PM_THRESHOLD_KEY_SUFFIXS:
                     key = self.convert_pm_prefix_to_threshold_prefix(
                         prefix) + suffix
-                    thresholds.append(
-                        float(sfp_threshold_dict[key]) if key in sfp_threshold_dict else None)
+                    threshold = None
+                    if key in sfp_threshold_dict and sfp_threshold_dict[key] != 'N/A':
+                        threshold = float(sfp_threshold_dict[key])
+                    thresholds.append(threshold)
 
                 tca_high, tca_low = None, None
                 if values[2] is not None and thresholds[0] is not None:
@@ -534,7 +583,10 @@ class SFPShow(object):
 
             output += tabulate(table,
                                ZR_PM_HEADER, disable_numparse=True).replace('\n', '\n' + indent)
-            output += '\n'
+            start_timestamp = self.beautify_pm_timestamp(
+                float(sfp_pm_dict.get('pm_win_start_time', None)))
+            start_timestamp_output = '\n{}PM window start time: {}'.format(indent, start_timestamp)
+            output = start_timestamp_output + output + '\n'
         else:
             output = ZR_PM_NOT_APPLICABLE_STR + '\n'
         return output
@@ -582,14 +634,14 @@ class SFPShow(object):
     def get_pm(self):
         if self.intf_name is not None:
             self.intf_pm[self.intf_name] = self.convert_interface_sfp_pm_to_cli_output_string(
-                self.db, self.intf_name)
+                self.db, self.intf_name, self.pm_window_size, self.pm_window_index)
         else:
             port_table_keys = self.db.keys(self.db.APPL_DB, "PORT_TABLE:*")
             for i in port_table_keys:
                 interface = re.split(':', i, maxsplit=1)[-1].strip()
                 if interface and interface.startswith(front_panel_prefix()) and not interface.startswith((backplane_prefix(), inband_prefix(), recirc_prefix())):
                     self.intf_pm[interface] = self.convert_interface_sfp_pm_to_cli_output_string(
-                        self.db, interface)
+                        self.db, interface, self.pm_window_size, self.pm_window_index)
 
     @multi_asic_util.run_on_multi_asic
     def get_status(self):
@@ -613,8 +665,10 @@ class SFPShow(object):
         click.echo(tabulate(sorted_port_table, header))
 
     def display_pm(self):
-        click.echo(
-            "\n".join([f"{k}: {v}" for k, v in natsorted(self.intf_pm.items())]))
+        output = self.beautify_pm_timestamp(time.time()) + '\n'
+        output += 'PM window: {}\n'.format(self.pm_window_size)
+        output += "\n".join([f"{k}: {v}" for k, v in natsorted(self.intf_pm.items())])
+        click.echo(output)
 
     def display_status(self):
         click.echo(
@@ -684,10 +738,26 @@ def presence(port, namespace):
 # 'pm' subcommand
 
 
-@cli.command()
-@click.option('-p', '--port', metavar='<port_name>', help="Display SFP PM for port <port_name> only")
+@cli.command(help="Display performance monitoring information")
+@click.option('-s', '--window_size',
+              type=click.Choice(['60sec', '15min', '24hrs']),
+              required=True,
+              help="Display data for the specified window size")
+@click.option('-i', '--window_index',
+              type=int,
+              required=True,
+              help="Display data for the specified window index")
+@click.option('-p', '--port', metavar='<port_name>', help="Display data for port <port_name> only")
 @click.option('-n', '--namespace', default=None, help="Display interfaces for specific namespace")
-def pm(port, namespace):
+def pm(window_size, window_index, port, namespace):
+     # Validate window_index based on window_size
+    if window_size == '60sec' and not 0 <= window_index <= 14:
+        raise click.BadArgumentUsage("For 60sec window size, window_index should be between 0 and 14.")
+    elif window_size == '15min' and not 0 <= window_index <= 11:
+        raise click.BadArgumentUsage("For 15min window size, window_index should be between 0 and 11.")
+    elif window_size == '24hrs' and not 0 <= window_index <= 1:
+        raise click.BadArgumentUsage("For 24hrs window size, window_index should be between 0 and 1.")
+
     if port and multi_asic.is_multi_asic() and namespace is None:
         try:
             namespace = multi_asic.get_namespace_for_port(port)
@@ -695,7 +765,7 @@ def pm(port, namespace):
             display_invalid_intf_pm(port)
             sys.exit(1)
 
-    sfp = SFPShow(port, namespace)
+    sfp = SFPShow(port, namespace, False, window_size, window_index)
     sfp.get_pm()
     sfp.display_pm()