--- a/scripts/sfpshow
+++ b/scripts/sfpshow
@@ -239,6 +239,9 @@ ZR_PM_INFO_MAP = {
     'EVM': ['%', 'evm']
 }
 
+ZR_PM_NOT_AVAILABLE_STR = \
+'Transceiver performance monitoring data not available for the requested window'
+
 ZR_PM_NOT_APPLICABLE_STR = 'Transceiver performance monitoring not applicable'
 
 QSFP_STATUS_NOT_APPLICABLE_STR = 'Transceiver status info not applicable'
@@ -513,7 +516,15 @@ class SFPShow(object):
 
     def get_pm_dict_for_specified_window(self, db, interface_name,
                                          window_size, window_index):
-        """Get PM dict for specified window size and index"""
+        """
+        Get PM dict for specified window size and index
+
+        Returns:
+            dict: PM dict for specified window size and index.
+                return None if PM data is not applicable for this interface;
+                return {} if PM data is applicable for this interface,
+                but not avaiable for this specified window.
+        """
         pm_windows_dict = db.get_all(
             db.STATE_DB, '{}|{}'.format(self.PM_TABLE_NAME_IN_DB, interface_name))
         if not pm_windows_dict:
@@ -527,7 +538,7 @@ class SFPShow(object):
             win_name = 'window' + str(i)
             if (win_name not in pm_windows_dict or
                 pm_windows_dict[win_name] == 'N/A'):
-                pm_dict = None
+                pm_dict = {}
             else:
                 # Convert string to dict
                 pm_dict = ast.literal_eval(pm_windows_dict[win_name])
@@ -537,7 +548,7 @@ class SFPShow(object):
                 pm_dict['pm_win_current'] == 'true'):
                 cur_win_idx = i
         if cur_win_idx == -1:
-            return None
+            return {}
         return pm_windows_for_win_size[(cur_win_idx -
                                         start_win_idx_in_db -
                                         window_index +
@@ -591,10 +602,12 @@ class SFPShow(object):
                 float(sfp_pm_dict.get('pm_win_end_time', None)))
             start_timestamp_output = '\n{}PM window start time: {}'.format(indent, start_timestamp)
             end_timestamp_output = '\n{}PM window end time:   {}'.format(indent, end_timestamp)
-            output = start_timestamp_output + end_timestamp_output + output + '\n'
+            output = start_timestamp_output + end_timestamp_output + output
+        elif sfp_pm_dict is None:
+            output = ZR_PM_NOT_APPLICABLE_STR
         else:
-            output = ZR_PM_NOT_APPLICABLE_STR + '\n'
-        return output
+            output = ZR_PM_NOT_AVAILABLE_STR
+        return output + '\n'
 
     @multi_asic_util.run_on_multi_asic
     def get_eeprom(self):