<!--
    Please make sure you've read and understood our contributing guidelines:
    https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

    ** Make sure all your commits include a signature generated with `git commit -s` **

    If this is a bug fix, make sure your description includes "closes #xxxx",
    "fixes #xxxx" or "resolves #xxxx" so that GitHub automatically closes the related
    issue when the PR is merged.

    If you are adding/modifying/removing any command or utility script, please also
    make sure to add/modify/remove any unit tests from the tests
    directory as appropriate.

    If you are modifying or removing an existing 'show', 'config' or 'sonic-clear'
    subcommand, or you are adding a new subcommand, please make sure you also
    update the Command Line Reference Guide (doc/Command-Reference.md) to reflect
    your changes.

    Please provide the following information:
-->

HLD can be found here: 
https://github.com/sonic-net/SONiC/pull/1258
Backend PRs:
https://github.com/sonic-net/sonic-platform-common/pull/387
https://github.com/sonic-net/sonic-platform-daemons/pull/390

#### What I did
According to window-based PM [HLD](https://github.com/sonic-net/SONiC/blob/d0364205764aa49decfca1eb669c76d7f0c8f4e0/doc/platform_api/CMIS_and_C-CMIS_support_for_ZR.md#74-cli-sample-output-format), add the advanced CLI for it.

#### How I did it
1. CLI ```show interfaces transceiver pm current <window_size{60sec/15min/24hrs}>``` will show the current PM data for the particular window_size (can be either 60sec/15min/24hrs)

2. CLI ```show interfaces transceiver pm history <window_size{60sec/15min/24hrs}> window <window_index{int}>``` will show the history PM data for the last Nth window(window_index) for the particular window_size.

#### How to verify it
With dummy pm DB data, verified CLI output with different combination of current/history window_size/window_index options.


#### Previous command output (if the output of a command-line utility has changed)
Today's pm CLI (```show int trans pm```) doesn't have current/history subcommand for window based data.
Output format is still the same.

#### New command output (if the output of a command-line utility has changed)

```
root@sonic:/home/cisco# show interfaces transceiver pm current 60sec Ethernet0 -n asic0                                                                                                                                                        
Wed Jul 05 18:59:59 UTC 2023                                                                                                                                                                                                      
PM window: 60sec                                                                                                                                                                                                                  
Ethernet0:                                                                                                                                                                                                                        
    PM window start time: Wed Jul 05 18:59:02 UTC 2023                                                                                                                                                                            
    Parameter        Unit    Min        Avg        Max        Threshold    Threshold    Threshold     Threshold    Threshold    Threshold                                                                                         
                                                              High         High         Crossing      Low          Low          Crossing                                                                                          
                                                              Alarm        Warning      Alert-High    Alarm        Warning      Alert-Low                                                                                         
    ---------------  ------  ---------  ---------  ---------  -----------  -----------  ------------  -----------  -----------  -----------                                                                                       
    Tx Power         dBm     0.0        0.0        0.0        -5.0         -6.0         True          -16.99       -16.003      False                                                                                             
    Rx Total Power   dBm     0.0        0.0        0.0        2.0          0.0          False         -21.0        -18.0        False                                                                                             
    Rx Signal Power  dBm     0.0        0.0        0.0        13.0         10.0         False         -18.0        -15.0        False                                                                                             
    CD-short link    ps/nm   6094848.0  6094848.0  6094848.0  1000.0       500.0        True          -1000.0      -500.0       False                                                                                             
    PDL              dB      0.3        0.5        0.8        4.0          4.0          False         0.0          0.0          False                                                                                             
    OSNR             dB      36.5       36.5       36.5       99.0         99.0         False         0.0          0.0          False                                                                                             
    eSNR             dB      30.5       30.5       30.5       99.0         99.0         False         0.0          0.0          False                                                                                             
    CFO              MHz     -66.0      34.9375    123.0      3800.0       3800.0       False         -3800.0      -3800.0      False                                                                                             
    DGD              ps      4.81       5.304375   5.68       7.0          7.0          False         0.0          0.0          False                                                                                             
    SOPMD            ps^2    0.0        0.0        0.0        655.35       655.35       False         0.0          0.0          False
    SOP ROC          krad/s  0.0        1.0        3.0        N/A          N/A          N/A           N/A          N/A          N/A
    Pre-FEC BER      N/A     7.99E-05   1.41E-04   1.67E-04   1.25E-02     1.10E-02     0.0           0.0          0.0          0.0
    Post-FEC BER     N/A     0.0        0.0        0.0        1000.0       1.0          False         0.0          0.0          False
    EVM              %       100.0      100.0      100.0      N/A          N/A          N/A           N/A          N/A          N/A


root@sonic:/home/cisco# show interfaces transceiver pm history 60sec window 1 Ethernet0 -n asic0
Wed Jul 05 18:59:59 UTC 2023   
PM window: 60sec
Ethernet0: 
    PM window start time: Wed Jul 05 18:57:54 UTC 2023
    Parameter        Unit    Min        Avg        Max        Threshold    Threshold    Threshold     Threshold    Threshold    Threshold
                                                              High         High         Crossing      Low          Low          Crossing
                                                              Alarm        Warning      Alert-High    Alarm        Warning      Alert-Low
    ---------------  ------  ---------  ---------  ---------  -----------  -----------  ------------  -----------  -----------  -----------
    Tx Power         dBm     0.0        0.0        0.0        -5.0         -6.0         True          -16.99       -16.003      False
    Rx Total Power   dBm     0.0        0.0        0.0        2.0          0.0          False         -21.0        -18.0        False
    Rx Signal Power  dBm     0.0        0.0        0.0        13.0         10.0         False         -18.0        -15.0        False
    CD-short link    ps/nm   6094848.0  6094848.0  6094848.0  1000.0       500.0        True          -1000.0      -500.0       False
    PDL              dB      0.3        0.50625    0.7        4.0          4.0          False         0.0          0.0          False
    OSNR             dB      36.5       36.5       36.5       99.0         99.0         False         0.0          0.0          False
    eSNR             dB      30.5       30.5       30.5       99.0         99.0         False         0.0          0.0          False
    CFO              MHz     -41.0      46.6875    152.0      3800.0       3800.0       False         -3800.0      -3800.0      False
    DGD              ps      4.87       5.2878125  5.75       7.0          7.0          False         0.0          0.0          False
    SOPMD            ps^2    0.0        0.0        0.0        655.35       655.35       False         0.0          0.0          False
    SOP ROC          krad/s  0.0        1.0        5.0        N/A          N/A          N/A           N/A          N/A          N/A
    Pre-FEC BER      N/A     9.94E-05   1.39E-04   2.24E-04   1.25E-02     1.10E-02     0.0           0.0          0.0          0.0
    Post-FEC BER     N/A     0.0        0.0        0.0        1000.0       1.0          False         0.0          0.0          False
    EVM              %       100.0      100.0      100.0      N/A          N/A          N/A           N/A          N/A          N/A


root@sonic:/home/cisco# show interfaces transceiver pm history 60sec window 2 Ethernet0 -n asic0
Wed Jul 05 18:59:59 UTC 2023   
PM window: 60sec
Ethernet0: 
    PM window start time: Wed Jul 05 18:56:47 UTC 2023
    Parameter        Unit    Min        Avg        Max        Threshold    Threshold    Threshold     Threshold    Threshold    Threshold
                                                              High         High         Crossing      Low          Low          Crossing
                                                              Alarm        Warning      Alert-High    Alarm        Warning      Alert-Low
    ---------------  ------  ---------  ---------  ---------  -----------  -----------  ------------  -----------  -----------  -----------
    Tx Power         dBm     0.0        0.0        0.0        -5.0         -6.0         True          -16.99       -16.003      False
    Rx Total Power   dBm     0.0        0.0        0.0        2.0          0.0          False         -21.0        -18.0        False
    Rx Signal Power  dBm     0.0        0.0        0.0        13.0         10.0         False         -18.0        -15.0        False
    CD-short link    ps/nm   6094848.0  6094848.0  6094848.0  1000.0       500.0        True          -1000.0      -500.0       False
    PDL              dB      0.4        0.590625   0.8        4.0          4.0          False         0.0          0.0          False
    OSNR             dB      36.5       36.5       36.5       99.0         99.0         False         0.0          0.0          False
    eSNR             dB      30.5       30.5       30.5       99.0         99.0         False         0.0          0.0          False
    CFO              MHz     -20.0      52.65625   131.0      3800.0       3800.0       False         -3800.0      -3800.0      False
    DGD              ps      4.87       5.2728125  5.81       7.0          7.0          False         0.0          0.0          False
    SOPMD            ps^2    0.0        0.0        0.0        655.35       655.35       False         0.0          0.0          False
    SOP ROC          krad/s  0.0        1.0        5.0        N/A          N/A          N/A           N/A          N/A          N/A
    Pre-FEC BER      N/A     9.44E-05   1.51E-04   2.51E-04   1.25E-02     1.10E-02     0.0           0.0          0.0          0.0
    Post-FEC BER     N/A     0.0        0.0        0.0        1000.0       1.0          False         0.0          0.0          False
    EVM              %       100.0      100.0      100.0      N/A          N/A          N/A           N/A          N/A          N/A



```